Don’t error on clean if no intermediaries
We check exists rather than isDirectory because we will get a better error if the path is a file (for example) instead and we try to rmdir it.

+++ Sources/swift-build/main.swift
@@ -70,7 +70,9 @@ do {
         usage()
 
     case .Clean(.Dist):
-        try rmtree(opts.path.Packages)
+        if opts.path.Packages.exists {
+            try rmtree(opts.path.Packages)
+        }
         fallthrough
 
     case .Clean(.Build):
@@ -86,7 +88,9 @@ do {
         let versionData = Path.join(opts.path.build, "versionData")
         if versionData.isDirectory { try rmtree(versionData) }
 
-        try rmdir(opts.path.build)
+        if opts.path.build.exists {
+            try rmdir(opts.path.build)
+        }
 
     case .Doctor:
         doctor()
